requirements.txt
streamlit==1.36.0
pillow==10.4.0
